requirements.txt
inquirer==3.4.0
prettytable==3.15.1
PyYAML==6.0.2
termcolor==2.5.0
